submission/HW2/110511233/requirements.txt
langchain
langchain-openai
python-dotenv
pandas